392/396

+++ ft_printf/libft/ft_pow.c
@@ -6,11 +6,11 @@
 /*   By: angavrel <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/27 19:38:26 by angavrel          #+#    #+#             */
-/*   Updated: 2017/01/27 20:07:31 by angavrel         ###   ########.fr       */
+/*   Updated: 2017/02/16 17:36:52 by angavrel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-long	ft_pow(long n, int pow)
+double	ft_pow(double n, int pow)
 {
 	return (pow ? n * ft_pow(n, pow - 1) : 1);
 }
