+++ ft_printf/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: angavrel <angavrel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/06 18:29:58 by angavrel          #+#    #+#             */
-/*   Updated: 2017/02/16 11:17:07 by angavrel         ###   ########.fr       */
+/*   Updated: 2017/02/16 17:34:42 by angavrel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -141,7 +141,7 @@ int					ft_abs(int n);
 float				ft_fclamp(float n, float min, float max);
 char				*ft_str2lower(char *s);
 char				*ft_itoa_base(int n, int b);
-long				ft_pow(long n, int pow);
+double				ft_pow(double n, int pow);
 
 /*
 ** printf functions
